try to use firefox in travis
