--- test/common/GebConfig.groovy
@@ -1,10 +1,0 @@
-import org.openqa.selenium.chrome.ChromeDriver
-
-driver = {
-    def driverInstance = new ChromeDriver()
-    driverInstance.manage().window().maximize()
-    driverInstance
-}
-
-baseNavigatorWaiting = true
-atCheckWaiting = true
