+++ test/functional/org/zenboot/portal/api/ZenbootApiSpec.groovy
@@ -4,7 +4,7 @@ import groovyx.net.http.RESTClient
 import spock.lang.Specification
 
 abstract class ZenbootApiSpec extends Specification {
-    def baseUrl = System.properties.get('grails.testing.functional.baseUrl')
+    def baseUrl = System.properties.get('grails.testing.functional.baseUrl') ?: 'http://localhost:8080/zenboot'
     def http
 
     def setup() {
